Extract transformVergunningFrontendV2 to abstract decos transformDecosZaakFrontend
