+++ src/server/services/vergunningen-v2/vergunningen.ts
@@ -1,22 +1,21 @@
 import memoizee from 'memoizee';
-import { generatePath } from 'react-router-dom';
-import slug from 'slugme';
 
-import { VergunningFrontendV2 } from './config-and-types';
-import { VergunningV2 } from './config-and-types';
+import { VergunningFrontendV2, VergunningV2 } from './config-and-types';
 import { decosZaakTransformers } from './decos-zaken';
 import { getStatusSteps } from './vergunningen-status-steps';
 import { AppRoute, AppRoutes } from '../../../universal/config/routes';
 import { apiSuccessResult } from '../../../universal/helpers/api';
-import { defaultDateFormat } from '../../../universal/helpers/date';
 import { AuthProfileAndToken } from '../../auth/auth-types';
 import { DEFAULT_API_CACHE_TTL_MS } from '../../config/source-api';
 import { encryptSessionIdWithRouteIdParam } from '../../helpers/encrypt-decrypt';
 import { BffEndpoints } from '../../routing/bff-routes';
 import { generateFullApiUrlBFF } from '../../routing/route-helpers';
-import { fetchDecosZaak, fetchDecosZaken } from '../decos/decos-service';
+import {
+  fetchDecosZaak,
+  fetchDecosZaken,
+  transformDecosZaakFrontend,
+} from '../decos/decos-service';
 import { DecosZaakDocument } from '../decos/decos-types';
-import { getStatusDate, isExpired, toDateFormatted } from '../decos/helpers';
 import { decryptEncryptedRouteParamAndValidateSessionID } from '../shared/decrypt-route-param';
 
 function transformVergunningFrontend(
@@ -24,50 +23,11 @@ function transformVergunningFrontend(
   vergunning: VergunningV2,
   appRoute: AppRoute
 ) {
-  const idEncrypted = encryptSessionIdWithRouteIdParam(
+  const vergunningFrontend = transformDecosZaakFrontend<VergunningV2>(
     sessionID,
-    vergunning.key
+    vergunning,
+    appRoute
   );
-  const vergunningFrontend: VergunningFrontendV2 = {
-    ...vergunning,
-    dateDecisionFormatted: toDateFormatted(vergunning.dateDecision),
-    dateInBehandeling: getStatusDate('In behandeling', vergunning),
-    dateInBehandelingFormatted: toDateFormatted(
-      getStatusDate('In behandeling', vergunning)
-    ),
-    dateRequestFormatted: defaultDateFormat(vergunning.dateRequest),
-    // Assign Status steps later on
-    steps: [],
-    // Adds an url with encrypted id to the BFF Detail page api for vergunningen.
-    fetchDocumentsUrl: generateFullApiUrlBFF(
-      BffEndpoints.VERGUNNINGENv2_DOCUMENTS_LIST,
-      [{ id: idEncrypted }]
-    ),
-    link: {
-      to: generatePath(appRoute, {
-        title: slug(vergunning.caseType, {
-          lower: true,
-        }),
-        id: vergunning.id,
-      }),
-      title: `Bekijk hoe het met uw aanvraag staat`,
-    },
-  };
-
-  // If a vergunning has both dateStart and dateEnd add formatted dates and an expiration indication.
-  if (
-    'dateEnd' in vergunning &&
-    'dateStart' in vergunning &&
-    vergunning.dateStart &&
-    vergunning.dateEnd
-  ) {
-    vergunningFrontend.isExpired = isExpired(vergunning);
-    vergunningFrontend.dateStartFormatted = defaultDateFormat(
-      vergunning.dateStart
-    );
-    vergunningFrontend.dateEndFormatted = defaultDateFormat(vergunning.dateEnd);
-  }
-
   // Assign the definitive status steps
   vergunningFrontend.steps = getStatusSteps(vergunningFrontend);
 
@@ -89,7 +49,7 @@ async function fetchVergunningenV2_(
     const decosVergunningen = response.content;
     const vergunningenFrontend: VergunningFrontendV2[] = decosVergunningen.map(
       (vergunning) =>
-        transformVergunningFrontend(
+        transformDecosZaakFrontend<VergunningV2>(
           authProfileAndToken.profile.sid,
           vergunning,
           appRoute
